+++ src/main/java/com/thealgorithms/stacks/DuplicateBrackets.java
@@ -1,4 +1,4 @@
-package com.thealgorithms.datastructures.stacks;
+package com.thealgorithms.stacks;
 
 //  1. You are given a string exp representing an expression.
 // 2. Assume that the expression is balanced  i.e. the opening and closing brackets match with each
