+++ src/test/java/com/thealgorithms/stacks/CalculateMaxOfMinTest.java
@@ -1,9 +1,8 @@
-package com.thealgorithms.others;
+package com.thealgorithms.stacks;
+
+import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
-
-import com.thealgorithms.datastructures.stacks.CalculateMaxOfMin;
-import org.junit.jupiter.api.Test;
 
 public class CalculateMaxOfMinTest {
 
@@ -11,48 +10,49 @@ public class CalculateMaxOfMinTest {
     void testForOneElement() {
         int[] a = {10, 20, 30, 50, 10, 70, 30};
         int k = CalculateMaxOfMin.calculateMaxOfMin(a);
-        assertTrue(k == 70);
+        assertEquals(70, k);
     }
 
     @Test
     void testForTwoElements() {
         int[] a = {5, 3, 2, 6, 3, 2, 6};
         int k = CalculateMaxOfMin.calculateMaxOfMin(a);
-        assertTrue(k == 6);
+        assertEquals(6, k);
     }
 
     @Test
     void testForThreeElements() {
         int[] a = {10, 10, 10, 10, 10, 10, 10};
         int k = CalculateMaxOfMin.calculateMaxOfMin(a);
-        assertTrue(k == 10);
+        assertEquals(10, k);
     }
 
     @Test
     void testForFourElements() {
         int[] a = {70, 60, 50, 40, 30, 20};
         int k = CalculateMaxOfMin.calculateMaxOfMin(a);
-        assertTrue(k == 70);
+        assertEquals(70, k);
     }
 
     @Test
     void testForFiveElements() {
         int[] a = {50};
         int k = CalculateMaxOfMin.calculateMaxOfMin(a);
-        assertTrue(k == 50);
+        assertEquals(50, k);
     }
 
     @Test
     void testForSixElements() {
         int[] a = {1, 4, 7, 9, 2, 4, 6};
         int k = CalculateMaxOfMin.calculateMaxOfMin(a);
-        assertTrue(k == 9);
+        assertEquals(9, k);
     }
 
     @Test
     void testForSevenElements() {
         int[] a = {-1, -5, -7, -9, -12, -14};
         int k = CalculateMaxOfMin.calculateMaxOfMin(a);
-        assertTrue(k == -1);
+        assertEquals(-1, k);
     }
+
 }
